+++ features/fixtures/mazerunner/app/src/main/java/com/bugsnag/android/mazerunner/MazerunnerApp.kt
@@ -10,6 +10,7 @@ class MazerunnerApp : Application() {
         super.onCreate()
         triggerStartupAnrIfRequired()
         setupNonSdkUsageStrictMode()
+        triggerManualSessionIfRequired()
     }
 
     /**