+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * FeatureFlags are now a copy-on-write structure, and so don't need to be defensive copied on a crashing thread
   [#2005](https://github.com/bugsnag/bugsnag-android/pull/2005)
 
+* Allow `Bugsnag.startSession` to be called with automatic session tracking, and not have the first manual session be over written by the first automatic session.
+  [#2006](https://github.com/bugsnag/bugsnag-android/pull/2006)
+
 ## 6.3.0 (2024-03-19)
 
 ### Enhancements